added navdata
